Adding visual tests for sky-paging component
Removing visual tests from list-paging since it adds nothing visual
