--- src/modules/list-paging/fixtures/list-paging.component.visual-fixture.html
@@ -1,3 +1,0 @@
-<div id="screenshot-list-paging">
-  <sky-list-paging pageSize="2" maxPages="3" pageNumber="1"></sky-list-paging>
-</div>
