+++ src/modules/paging/fixtures/paging.component.visual-fixture.html
@@ -1,0 +1,3 @@
+<div id="screenshot-list-paging">
+  <sky-paging pageSize="2" maxPages="3" currentPage="1" itemCount="7"></sky-paging>
+</div>
