+++ src/modules/paging/fixtures/paging.component.visual-fixture.ts
@@ -1,0 +1,35 @@
+import {
+  Component,
+  NgModule
+} from '@angular/core';
+import { BrowserModule } from '@angular/platform-browser';
+
+import {
+  SkyModule
+} from '../../../../src/core';
+
+import { Bootstrapper } from '../../../../visual/bootstrapper';
+
+@Component({
+  selector: 'sky-demo-app',
+  templateUrl: './paging.component.visual-fixture.html'
+})
+export class AppComponent {
+
+}
+
+@NgModule({
+  imports: [
+    BrowserModule,
+    SkyModule
+  ],
+  declarations: [
+    AppComponent
+  ],
+  bootstrap: [
+    AppComponent
+  ]
+})
+class AppModule { }
+
+Bootstrapper.bootstrapModule(AppModule);
